[+] width all be careful %g i go sleep tommorow i test him

+++ ressources/my_intlen.c
@@ -6,7 +6,7 @@
 */
 #include "my_printf.h"
 
-int my_intlen(int nb)
+int my_intlen(long nb)
 {
     int compt = 0;
